split cuda reoutines in a new module
